Simplify VisionAPI by removing ContextAPI baseclass
